# back/app/app/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "`NFTicket` 프로젝트 백엔드 API 문서",
    "version": "1.0.0",
    "title": "NFTicket API Documentation",
    "termsOfService": "https://nfticket.plus",
    "contact": {
      "email": "nfticket@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "nfticket.plus",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/sale": {
      "post": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 정보 생성",
        "description": "",
        "operationId": "createSale",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "생성할 티켓 판매 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Sale"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "성공",
            "schema": {
              "type": "object",
              "properties": {
                "sale_id": {
                  "type": "integer",
                  "format": "int64",
                  "description": "새로 생성된 티켓 판매 정보 ID"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Invalid request"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "get": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 정보 전체 반환",
        "description": "",
        "operationId": "getSale",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sale"
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/sale/{sale_id}": {
      "get": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 정보 반환",
        "description": "",
        "operationId": "getSale",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "$ref": "#/definitions/Sale"
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Role doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "patch": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 정보 수정",
        "description": "",
        "operationId": "editSale",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "갱신할 티켓 판매 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Sale"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Invalid request"
                }
              }
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/sale/{sale_id}/show-schedule-id": {
      "get": {
        "tags": [
          "Sale"
        ],
        "summary": "공연 스케줄 ID 정보 반환",
        "description": "",
        "operationId": "getSaleShowScheduleId",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "$ref": "#/definitions/SaleShowScheduleId"
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "patch": {
        "tags": [
          "Sale"
        ],
        "summary": "공연 스케줄 ID 정보 수정",
        "description": "",
        "operationId": "editSaleShowScheduleId",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "갱신할 공연 스케줄 ID 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SaleShowScheduleId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Invalid request"
                }
              }
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/sale/{sale_id}/description": {
      "get": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 설명 정보 반환",
        "description": "",
        "operationId": "getSaleDescription",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "$ref": "#/definitions/SaleDescription"
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "patch": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 설명 정보 수정",
        "description": "",
        "operationId": "editSaleDescription",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "갱신할 티켓 판매 설명 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SaleDescription"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Invalid request"
                }
              }
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/sale/{sale_id}/started-at": {
      "get": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 시작 시간 반환",
        "description": "",
        "operationId": "getSaleStartedAt",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "$ref": "#/definitions/SaleStartedAt"
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "patch": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 시작 시간 수정",
        "description": "",
        "operationId": "editSaleStartedAt",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "갱신할 티켓 판매 시작 시간 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SaleStartedAt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Invalid request"
                }
              }
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/sale/{sale_id}/ended-at": {
      "get": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 종료 시간 반환",
        "description": "",
        "operationId": "getSaleEndedAt",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "$ref": "#/definitions/SaleEndedAt"
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "patch": {
        "tags": [
          "Sale"
        ],
        "summary": "티켓 판매 종료 시간 수정",
        "description": "",
        "operationId": "editSaleEndedAt",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sale_id",
            "description": "티켓 판매 정보 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "갱신할 티켓 판매 종료 시간 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SaleEndedAt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Invalid request"
                }
              }
            }
          },
          "404": {
            "description": "티켓 판매 정보 없음",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "format": "string",
                  "description": "Sale doesn't exist"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SaleShowScheduleId": {
      "type": "object",
      "description": "공연 스케줄 ID",
      "properties": {
        "show_schedule_id": {
          "description": "공연 스케줄 ID",
          "type": "integer",
          "format": "int256",
          "example": 123
        }
      }
    },
    "SaleDescription": {
      "type": "object",
      "description": "티켓 판매에 관한 설명",
      "properties": {
        "description": {
          "description": "티켓 판매에 관한 설명",
          "type": "string",
          "format": "string",
          "example": "지킬앤하이드는..."
        }
      }
    },
    "SaleStartedAt": {
      "type": "object",
      "description": "티켓 판매 시작 시간",
      "properties": {
        "started_at": {
          "description": "티켓 판매 시작 시간",
          "type": "string",
          "format": "datetime",
          "example": "2022-03-30T17:24:14.000Z"
        }
      }
    },
    "SaleEndedAt": {
      "type": "object",
      "description": "티켓 판매 종료 시간",
      "properties": {
        "ended_at": {
          "description": "티켓 판매 종료 시간",
          "type": "string",
          "format": "datetime",
          "example": "2022-03-30T17:24:14.000Z"
        }
      }
    },
    "Sale": {
      "type": "object",
      "description": "티켓 판매 정보",
      "required": [
        "show_schedule_id",
        "description",
        "started_at",
        "ended_at"
      ],
      "properties": {
        "show_schedule_id": {
          "description": "공연 스케줄 ID",
          "type": "integer",
          "format": "int256",
          "example": 123
        },
        "description": {
          "description": "티켓 판매에 관한 설명",
          "type": "string",
          "format": "string",
          "example": "지킬앤하이드는..."
        },
        "started_at": {
          "description": "티켓 판매 시작 시간",
          "type": "string",
          "format": "datetime",
          "example": "2022-03-30T17:24:14.000Z"
        },
        "ended_at": {
          "description": "티켓 판매 종료 시간",
          "type": "string",
          "format": "datetime",
          "example": "2022-03-30T17:24:14.000Z"
        }
      }
    },
...
  },
...
}